WIP: do not use puppeteer, share a code between a server and a client
